refactor socket
